Add files via upload
Modified translate.py: added auto_decode() function to read a txt file with input sentences (.en) and writes the output sentences generated with the trained model in another file (_out_gru.txt or _out_mtgru.txt). This generated file is then compared with the file containing the expected sentences (.fr) to get the BLEU and ROUGE calculations (and in the future also METEOR).